move quantizers to a new quantizers.py from types.py
