+++ hls4ml/converters/keras/core.py
@@ -1,5 +1,6 @@
 from hls4ml.converters.keras_to_hls import get_weights_data, keras_handler, parse_default_keras_layer
-from hls4ml.model.types import BinaryQuantizer, IntegerPrecisionType, TernaryQuantizer
+from hls4ml.model.quantizers import BinaryQuantizer, TernaryQuantizer
+from hls4ml.model.types import IntegerPrecisionType
 
 
 @keras_handler('InputLayer')
